refactor: transactions table
