+++ ui/portal/web/src/components/explorer/BlockExplorer.tsx
@@ -24,6 +24,7 @@ import { useNavigate } from "@tanstack/react-router";
 import type React from "react";
 import type { PropsWithChildren } from "react";
 import { HeaderExplorer } from "./HeaderExplorer";
+import { TransactionsTable } from "./TransactionsTable";
 
 type BlockExplorerProps = {
   height: string;
@@ -304,64 +305,13 @@ const BlockNotFound: React.FC = () => {
 };
 
 const BlockTable: React.FC = () => {
-  const navigate = useNavigate();
   const { data } = useBlockExplorer();
 
   if (!data?.searchBlock) return null;
 
   const { transactions } = data.searchBlock;
 
-  const columns: TableColumn<IndexedTransaction> = [
-    {
-      header: "Hash",
-      cell: ({ row }) => (
-        <Cell.TxHash
-          hash={row.original.hash}
-          navigate={() => navigate({ to: `/tx/${row.original.hash}` })}
-        />
-      ),
-    },
-    {
-      header: "Block",
-      cell: ({ row }) => (
-        <Cell.BlockHeight
-          blockHeight={row.original.blockHeight}
-          navigate={() => navigate({ to: `/block/${row.original.blockHeight}` })}
-        />
-      ),
-    },
-    {
-      header: "Age",
-      cell: ({ row }) => <Cell.Age date={row.original.createdAt} addSuffix />,
-    },
-    {
-      header: "Sender",
-      cell: ({ row }) => (
-        <Cell.Sender sender={row.original.sender} navigate={(url) => navigate({ to: url })} />
-      ),
-    },
-    {
-      header: "Actions",
-      cell: ({ row }) => <Cell.TxMessages messages={row.original.messages} />,
-    },
-    {
-      header: "Result",
-      cell: ({ row }) => {
-        const { hasSucceeded, messages } = row.original;
-
-        return (
-          <Cell.TxResult
-            className="justify-end"
-            isSuccess={hasSucceeded}
-            text={m["explorer.txs.result"]({ result: String(hasSucceeded) })}
-            total={messages.length}
-          />
-        );
-      },
-    },
-  ];
-
-  return transactions.length ? <Table data={transactions} columns={columns} /> : null;
+  return <TransactionsTable transactions={transactions} />;
 };
 
 export const BlockExplorer = Object.assign(BlockContainer, {
